рефакторинг - переименование репозиториев
