+++ hw7/src/main/java/ru/otus/hw/repositories/AuthorRepositoryJdbc.java
@@ -1,7 +1,6 @@
 package ru.otus.hw.repositories;
 
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
@@ -9,13 +8,12 @@ import ru.otus.hw.models.Author;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 @AllArgsConstructor
-public class JdbcAuthorRepository implements AuthorRepository {
+public class AuthorRepositoryJdbc implements AuthorRepository {
 
     final private JdbcTemplate jdbcTemplate;
