+++ hw7/src/main/java/ru/otus/hw/repositories/AuthorRepositoryJpa.java
@@ -1,0 +1,34 @@
+package ru.otus.hw.repositories;
+
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Repository;
+import ru.otus.hw.models.Author;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+@Primary
+public class AuthorRepositoryJpa implements AuthorRepository {
+
+    @PersistenceContext
+    private final EntityManager em;
+
+    public AuthorRepositoryJpa(EntityManager em) {
+        this.em = em;
+    }
+
+    @Override
+    public List<Author> findAll() {
+        var query = em.createQuery("select a from Author a", Author.class);
+        return query.getResultList();
+    }
+
+    @Override
+    public Optional<Author> findById(long id) {
+        return Optional.ofNullable(em.find(Author.class, id));
+    }
+
+}
