+++ hw7/src/main/java/ru/otus/hw/repositories/GenreRepositoryJdbc.java
@@ -8,13 +8,12 @@ import ru.otus.hw.models.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 @AllArgsConstructor
-public class JdbcGenreRepository implements GenreRepository {
+public class GenreRepositoryJdbc implements GenreRepository {
 
     final private JdbcTemplate jdbcTemplate;
     @Override
