+++ hw7/src/main/java/ru/otus/hw/repositories/GenreRepositoryJpa.java
@@ -4,7 +4,6 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
-import ru.otus.hw.models.Author;
 import ru.otus.hw.models.Genre;
 
 import java.util.List;
@@ -12,24 +11,23 @@ import java.util.Optional;
 
 @Repository
 @Primary
-public class JpaAuthorRepository implements AuthorRepository {
+public class GenreRepositoryJpa implements GenreRepository {
 
     @PersistenceContext
     private final EntityManager em;
 
-    public JpaAuthorRepository(EntityManager em) {
+    public GenreRepositoryJpa(EntityManager em) {
         this.em = em;
     }
 
     @Override
-    public List<Author> findAll() {
-        var query = em.createQuery("select a from Author a", Author.class);
+    public List<Genre> findAll() {
+        var query = em.createQuery("select g from Genre g", Genre.class);
         return query.getResultList();
     }
 
     @Override
-    public Optional<Author> findById(long id) {
-        return Optional.ofNullable(em.find(Author.class, id));
+    public Optional<Genre> findById(long id) {
+        return Optional.ofNullable(em.find(Genre.class, id));
     }
-
 }
