--- hw7/src/main/java/ru/otus/hw/repositories/JpaGenreRepository.java
@@ -1,37 +1,0 @@
-package ru.otus.hw.repositories;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import lombok.AllArgsConstructor;
-import org.springframework.context.annotation.Primary;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import ru.otus.hw.models.Genre;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-@Primary
-public class JpaGenreRepository implements GenreRepository {
-
-    @PersistenceContext
-    private final EntityManager em;
-
-    public JpaGenreRepository(EntityManager em) {
-        this.em = em;
-    }
-
-    @Override
-    public List<Genre> findAll() {
-        var query = em.createQuery("select g from Genre g", Genre.class);
-        return query.getResultList();
-    }
-
-    @Override
-    public Optional<Genre> findById(long id) {
-        return Optional.ofNullable(em.find(Genre.class, id));
-    }
-}
