+++ hw7/src/main/java/ru/otus/hw/services/GenreServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.otus.hw.services;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.hw.models.Author;
@@ -13,6 +14,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Service
 @Transactional
+@Primary
 public class GenreServiceImpl implements GenreService {
     private final GenreRepository genreRepository;
